+++ src/basins/basins_utilities.jl
@@ -2,16 +2,17 @@ export basin_fractions, match_attractors!
 
 """
     basin_fractions(basins::Array) → fs::Dict
-Calculate the state space fraction of the basins of attraction encoded in `basins`. 
-The elements of `basins` are integers, enumerating the attractor that the entry of 
-`basins` converges to (i.e. like the output of [`basins_of_attraction`](@ref)).
+Calculate the state space fraction of the basins of attraction encoded in `basins`.
+The elements of `basins` are integers, enumerating the attractor that the entry of
+`basins` converges to (i.e., like the output of [`basins_of_attraction`](@ref)).
 Return a dictionary that maps attractor IDs to their relative fractions.
 
 In[^Menck2013] the authors use these fractions to quantify the stability of a basin of
 attraction, and specifically how it changes when a parameter is changed.
 
-[^Menck2013]: Menck, Heitzig, Marwan & Kurths. How basin stability complements the linear
-stability paradigm. [Nature Physics, 9(2), 89–92](https://doi.org/10.1038/nphys2516)
+[^Menck2013]:
+    Menck, Heitzig, Marwan & Kurths. How basin stability complements the linear
+    stability paradigm. [Nature Physics, 9(2), 89–92](https://doi.org/10.1038/nphys2516)
 """
 function basin_fractions(basins::AbstractArray)
     fs = Dict{eltype(basins), Float64}()
@@ -28,7 +29,7 @@ end
     match_attractors!(b₋, a₋, b₊, a₊, [, method = :distance])
 Attempt to match the attractors in basins/attractors `b₊, a₊` with those at `b₋, a₋`.
 `b` is an array whose values encode the attractor ID, while `a` is a dictionary mapping
-IDs to `Dataset`s containing the attractors (e.g. output of [`basins_of_attraction`](@ref)).
+IDs to `Dataset`s containing the attractors (i.e, output of [`basins_of_attraction`](@ref)).
 Typically the +,- mean after and before some change of parameter for a system.
 
 In [`basins_of_attraction`](@ref) different attractors get assigned different IDs, however